VOCPart/requirements.txt
kornia==0.2.0
torch==1.4.0
torchvision==0.5.0
numpy==1.17.3
tqdm==4.41.1
sklearn==0.0
scipy==1.4.1
matplotlib==3.1.1
tensorboard==2.0.1
Pillow==6.2.1